Only download compat-table if it's not already downloaded

+++ packages/babel-compat-data/scripts/download-compat-table.sh
@@ -2,6 +2,12 @@
 set -e
 
 COMPAT_TABLE_COMMIT=4e9369a699b0e15ba5c21586ce3bdd34299db9c1
+GIT_HEAD=build/compat-table/.git/HEAD
+
+if [ -f "$GIT_HEAD" -a "$(cat $GIT_HEAD)" == $COMPAT_TABLE_COMMIT ]; then
+  exit 0
+fi
+
 rm -rf build/compat-table
 mkdir -p build
 git clone --branch=gh-pages --single-branch --shallow-since=2019-11-14 https://github.com/kangax/compat-table.git build/compat-table
